completed unit testing
completed unit testing for the Open Telemetry
