+++ AutomatedTesting/AutomatedTesting.csproj
@@ -14,8 +14,8 @@
 	  <PackageReference Include="Microsoft.Extensions.Diagnostics.Testing" Version="8.10.0" />
 	  <PackageReference Include="Moq" Version="4.20.72" />
 	  <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
-	  <PackageReference Include="MSTest.TestAdapter" Version="3.8.0" />
-	  <PackageReference Include="MSTest.TestFramework" Version="3.8.0" />
+	  <PackageReference Include="MSTest.TestAdapter" Version="3.8.3" />
+	  <PackageReference Include="MSTest.TestFramework" Version="3.8.3" />
 	  <PackageReference Include="coverlet.collector" Version="6.0.4">
 		  <PrivateAssets>all</PrivateAssets>
 		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
