+++ AutomatedTesting/ConnectionTests/MappedService/QueryInboxTests.cs
@@ -516,7 +516,9 @@ namespace AutomatedTesting.ConnectionTests.MappedService
         }
 
         [TestMethod]
-        public async Task TestQueryAsyncWithTelemetryDataWithoutLinking()
+        [DataRow(false)]
+        [DataRow(true)]
+        public async Task TestQueryAsyncWithTelemetryData(bool withLinking)
         {
             #region Arrange
             (var listener, var capturedActivities, var sourceName) = ConnectionHelper.SetupTelemetry();
@@ -565,7 +567,7 @@ namespace AutomatedTesting.ConnectionTests.MappedService
 
             var contractConnection = ContractConnection.MappedServiceInstance()
                 .RegisterServiceConnection((props) => true, ServiceName, serviceConnection.Object)
-                .EnableOpenTelemetry(activitySource: sourceName, linkActivitiesAcrossSystems: false);
+                .EnableOpenTelemetry(activitySource: sourceName, linkActivitiesAcrossSystems: withLinking);
             #endregion
 
             #region Act
@@ -584,7 +586,7 @@ namespace AutomatedTesting.ConnectionTests.MappedService
             Assert.IsFalse(result.IsError);
             Assert.AreEqual(typeof(BasicQueryMessage).GetCustomAttribute<MessageChannelAttribute>(false)?.Name, messages[0].Channel);
             Assert.AreEqual(1, messageIDs.Count);
-            Assert.AreEqual(0, messages[0].Header.Keys.Count());
+            Assert.AreEqual((withLinking ? 2 : 0), messages[0].Header.Keys.Count());
             Assert.AreEqual("U-BasicQueryMessage-0.0.0.0", messages[0].MessageTypeID);
             Assert.IsTrue(messages[0].Data.Length > 0);
             Assert.AreEqual(testMessage, await JsonSerializer.DeserializeAsync<BasicQueryMessage>(new MemoryStream(messages[0].Data.ToArray())));
@@ -596,7 +598,7 @@ namespace AutomatedTesting.ConnectionTests.MappedService
                 "MQContract.PublishQueryMessage",
                 serviceConnection.Object.GetType(),
                 true,
-                false,
+                withLinking,
                 connectionName:ServiceName
             );
             ConnectionHelper.ValidateConsumeActivity<BasicResponseMessage>(
@@ -606,107 +608,6 @@ namespace AutomatedTesting.ConnectionTests.MappedService
                 serviceConnection.Object.GetType(),
                 true,
                 connectionName:ServiceName
-            );
-            #endregion
-
-            #region Verify
-            serviceConnection.Verify(x => x.QueryAsync(It.IsAny<ServiceMessage>(), It.IsAny<Guid>(), It.IsAny<CancellationToken>()), Times.Once);
-            serviceConnection.Verify(x => x.EstablishInboxSubscriptionAsync(It.IsAny<Action<ReceivedInboxServiceMessage>>(), It.IsAny<CancellationToken>()), Times.Once);
-            mockSubscription.Verify(x => x.EndAsync(), Times.Once);
-            #endregion
-        }
-
-        [TestMethod]
-        public async Task TestQueryAsyncWithTelemetryDataWithLinking()
-        {
-            #region Arrange
-            (var listener, var capturedActivities, var sourceName) = ConnectionHelper.SetupTelemetry();
-            var testMessage = new BasicQueryMessage("testMessage");
-            var responseMessage = new BasicResponseMessage("testResponse");
-            using var ms = new MemoryStream();
-            await JsonSerializer.SerializeAsync(ms, responseMessage);
-            var responseData = (ReadOnlyMemory<byte>)ms.ToArray();
-
-            var queryResult = new ServiceQueryResult(Guid.NewGuid().ToString(), new MessageHeader([]), "U-BasicResponseMessage-0.0.0.0", responseData);
-
-            var mockSubscription = new Mock<IServiceSubscription>();
-
-            var defaultTimeout = TimeSpan.FromMinutes(1);
-
-            List<Action<ReceivedInboxServiceMessage>> receivedActions = [];
-            List<ServiceMessage> messages = [];
-            List<Guid> messageIDs = [];
-            var acknowledgeCount = 0;
-
-
-            var serviceConnection = new Mock<IInboxQueryableMessageServiceConnection>();
-            serviceConnection.Setup(x => x.EstablishInboxSubscriptionAsync(Capture.In(receivedActions), It.IsAny<CancellationToken>()))
-                .ReturnsAsync(mockSubscription.Object);
-            serviceConnection.Setup(x => x.QueryAsync(Capture.In(messages), Capture.In(messageIDs), It.IsAny<CancellationToken>()))
-                .Returns((ServiceMessage message, Guid messageID, CancellationToken cancellationToken) =>
-                {
-                    foreach (var action in receivedActions)
-                        action(new(
-                            queryResult.ID,
-                            queryResult.MessageTypeID,
-                            message.Channel,
-                            queryResult.Header,
-                            messageID,
-                            queryResult.Data,
-                            () =>
-                            {
-                                acknowledgeCount++;
-                                return ValueTask.CompletedTask;
-                            }
-                        ));
-                    return ValueTask.FromResult(new TransmissionResult(message.ID));
-                });
-            serviceConnection.Setup(x => x.DefaultTimeout)
-                .Returns(defaultTimeout);
-
-            var contractConnection = ContractConnection.MappedServiceInstance()
-                .RegisterServiceConnection((props) => true, ServiceName, serviceConnection.Object)
-                .EnableOpenTelemetry(activitySource: sourceName, linkActivitiesAcrossSystems: true);
-            #endregion
-
-            #region Act
-            var stopwatch = Stopwatch.StartNew();
-            var result = await contractConnection.QueryAsync<BasicQueryMessage, BasicResponseMessage>(testMessage);
-            stopwatch.Stop();
-            Trace.WriteLine($"Time to publish message {stopwatch.ElapsedMilliseconds}ms");
-            await contractConnection.CloseAsync();
-            #endregion
-
-            #region Assert
-            Assert.IsTrue(await Helper.WaitForCount(messages, 1, TimeSpan.FromMinutes(1)));
-            Assert.IsNotNull(result);
-            Assert.AreEqual(queryResult.ID, result.ID);
-            Assert.IsNull(result.Error);
-            Assert.IsFalse(result.IsError);
-            Assert.AreEqual(typeof(BasicQueryMessage).GetCustomAttribute<MessageChannelAttribute>(false)?.Name, messages[0].Channel);
-            Assert.AreEqual(1, messageIDs.Count);
-            Assert.AreEqual(2, messages[0].Header.Keys.Count());
-            Assert.AreEqual("U-BasicQueryMessage-0.0.0.0", messages[0].MessageTypeID);
-            Assert.IsTrue(messages[0].Data.Length > 0);
-            Assert.AreEqual(testMessage, await JsonSerializer.DeserializeAsync<BasicQueryMessage>(new MemoryStream(messages[0].Data.ToArray())));
-            Assert.AreEqual(responseMessage, result.Result);
-            Assert.AreEqual(2, capturedActivities.Count);
-            ConnectionHelper.ValidatePublishActivity<BasicQueryMessage>(
-                messages[0],
-                capturedActivities[0],
-                "MQContract.PublishQueryMessage",
-                serviceConnection.Object.GetType(),
-                true,
-                true,
-                connectionName: ServiceName
-            );
-            ConnectionHelper.ValidateConsumeActivity<BasicResponseMessage>(
-                queryResult,
-                capturedActivities[1],
-                "MQContract.ConsumeQueryResponse",
-                serviceConnection.Object.GetType(),
-                true,
-                connectionName: ServiceName
             );
             #endregion
