+++ AutomatedTesting/ConnectionTests/SingleService/QueryTests.cs
@@ -1072,7 +1072,9 @@ namespace AutomatedTesting.ConnectionTests.SingleService
         }
 
         [TestMethod]
-        public async Task TestQueryAsyncWithTelemetryDataWithoutLinking()
+        [DataRow(false)]
+        [DataRow(true)]
+        public async Task TestQueryAsyncWithTelemetryData(bool withLinking)
         {
             #region Arrange
             (var listener, var capturedActivities, var sourceName) = ConnectionHelper.SetupTelemetry();
@@ -1097,7 +1099,7 @@ namespace AutomatedTesting.ConnectionTests.SingleService
                 .Returns(defaultTimeout);
 
             var contractConnection = ContractConnection.Instance(serviceConnection.Object)
-                .EnableOpenTelemetry(activitySource:sourceName,linkActivitiesAcrossSystems:false);
+                .EnableOpenTelemetry(activitySource:sourceName,linkActivitiesAcrossSystems:withLinking);
             #endregion
 
             #region Act
@@ -1116,7 +1118,7 @@ namespace AutomatedTesting.ConnectionTests.SingleService
             Assert.AreEqual(typeof(BasicQueryMessage).GetCustomAttribute<MessageChannelAttribute>(false)?.Name, messages[0].Channel);
             Assert.AreEqual(1, timeouts.Count);
             Assert.AreEqual(defaultTimeout, timeouts[0]);
-            Assert.AreEqual(0, messages[0].Header.Keys.Count());
+            Assert.AreEqual((withLinking ? 2 :0), messages[0].Header.Keys.Count());
             Assert.AreEqual("U-BasicQueryMessage-0.0.0.0", messages[0].MessageTypeID);
             Assert.IsTrue(messages[0].Data.Length > 0);
             Assert.AreEqual(testMessage, await JsonSerializer.DeserializeAsync<BasicQueryMessage>(new MemoryStream(messages[0].Data.ToArray())));
@@ -1128,81 +1130,7 @@ namespace AutomatedTesting.ConnectionTests.SingleService
                 "MQContract.PublishQueryMessage",
                 serviceConnection.Object.GetType(),
                 true,
-                false,
-                includePublish:false
-            );
-            ConnectionHelper.ValidateConsumeActivity<BasicResponseMessage>(
-                queryResult,
-                capturedActivities[1],
-                "MQContract.ConsumeQueryResponse",
-                serviceConnection.Object.GetType(),
-                true
-            );
-            #endregion
-
-            #region Verify
-            serviceConnection.Verify(x => x.QueryAsync(It.IsAny<ServiceMessage>(), It.IsAny<TimeSpan>(), It.IsAny<CancellationToken>()), Times.Once);
-            #endregion
-        }
-
-        [TestMethod]
-        public async Task TestQueryAsyncWithTelemetryDataWithLinking()
-        {
-            #region Arrange
-            (var listener, var capturedActivities, var sourceName) = ConnectionHelper.SetupTelemetry();
-            var testMessage = new BasicQueryMessage("testMessage");
-            var responseMessage = new BasicResponseMessage("testResponse");
-            using var ms = new MemoryStream();
-            await JsonSerializer.SerializeAsync(ms, responseMessage);
-            var responseData = (ReadOnlyMemory<byte>)ms.ToArray();
-
-            var queryResult = new ServiceQueryResult(Guid.NewGuid().ToString(), new MessageHeader([]), "U-BasicResponseMessage-0.0.0.0", responseData);
-
-
-            var defaultTimeout = TimeSpan.FromMinutes(1);
-
-            List<ServiceMessage> messages = [];
-            List<TimeSpan> timeouts = [];
-
-            var serviceConnection = new Mock<IQueryResponseMessageServiceConnection>();
-            serviceConnection.Setup(x => x.QueryAsync(Capture.In(messages), Capture.In(timeouts), It.IsAny<CancellationToken>()))
-                .ReturnsAsync(queryResult);
-            serviceConnection.Setup(x => x.DefaultTimeout)
-                .Returns(defaultTimeout);
-
-            var contractConnection = ContractConnection.Instance(serviceConnection.Object)
-                .EnableOpenTelemetry(activitySource: sourceName, linkActivitiesAcrossSystems: true);
-            #endregion
-
-            #region Act
-            var stopwatch = Stopwatch.StartNew();
-            var result = await contractConnection.QueryAsync<BasicQueryMessage, BasicResponseMessage>(testMessage);
-            stopwatch.Stop();
-            Trace.WriteLine($"Time to publish message {stopwatch.ElapsedMilliseconds}ms");
-            #endregion
-
-            #region Assert
-            Assert.IsTrue(await Helper.WaitForCount(messages, 1, TimeSpan.FromMinutes(1)));
-            Assert.IsNotNull(result);
-            Assert.AreEqual(queryResult.ID, result.ID);
-            Assert.IsNull(result.Error);
-            Assert.IsFalse(result.IsError);
-            Assert.AreEqual(typeof(BasicQueryMessage).GetCustomAttribute<MessageChannelAttribute>(false)?.Name, messages[0].Channel);
-            Assert.AreEqual(1, timeouts.Count);
-            Assert.AreEqual(defaultTimeout, timeouts[0]);
-            Assert.AreEqual(2, messages[0].Header.Keys.Count());
-            Assert.AreEqual("U-BasicQueryMessage-0.0.0.0", messages[0].MessageTypeID);
-            Assert.IsTrue(messages[0].Data.Length > 0);
-            Assert.AreEqual(testMessage, await JsonSerializer.DeserializeAsync<BasicQueryMessage>(new MemoryStream(messages[0].Data.ToArray())));
-            Assert.AreEqual(responseMessage, result.Result);
-            Assert.AreEqual(2, capturedActivities.Count);
-            ConnectionHelper.ValidatePublishActivity<BasicQueryMessage>(
-                messages[0],
-                capturedActivities[0],
-                "MQContract.PublishQueryMessage",
-                serviceConnection.Object.GetType(),
-                true,
-                true,
+                withLinking,
                 includePublish:false
             );
             ConnectionHelper.ValidateConsumeActivity<BasicResponseMessage>(
